use listeners for configuration
