+++ packages/server/customer-os-common-module/services/agent_capability/capability_add_meeting_notes_to_company.go
@@ -73,7 +73,7 @@ func (c *AddMeetingNotesToCompanyCapability) ValidateInput(input AddMeetingNotes
 	return nil
 }
 
-func (c *AddMeetingNotesToCompanyCapability) ValidateConfig(config postgres_entity.NoConfig) error {
+func (c *AddMeetingNotesToCompanyCapability) ValidateConfig(postgres_entity.NoConfig) error {
 	return nil
 }
